rename package to pydatasummary
