+++ Makefile
@@ -7,4 +7,4 @@ hooks-all:
 	poetry run pre-commit run --all-files
 
 pytest-cov:
-	poetry run pytest --cov-report term --cov=datasummary tests/
+	poetry run pytest --cov-report term --cov=pydatasummary tests/
