+++ src/pydatasummary/__init__.py
@@ -1,10 +1,10 @@
-# Define datasummary version
+# Define pydatasummary version
 from importlib_metadata import version as _v
 
 # __version__ = "0.0.1"
-__version__ = _v("datasummary")
+__version__ = _v("pydatasummary")
 
 del _v
 
-# Import datasummary objects
+# Import pydatasummary objects
 from .ds import *  # noqa: F401, F403, E402
